+++ packages/template-retail-react-app/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Show Bonus Product Label on OrderSummary component [#2524](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2524)
 - Added support for Shop in Store Functionality
 - Added support for PDP page to support Pickup in Store[#2537] (https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2537)
+- Compute store level inventory and refactor product view [#2553](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2553)
 - Show Bonus Product Label on OrderSummary component [#2524](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2524)